Added files via upload
I wanted to use `rvest` to scrape data from indeed. However, after spending a few days of trying, I couldn't get the scraping function to work, since indeed is a job search aggregator and the links it provides go to different company websites that have different structure. This is my first time scraping so I also didn't know enough R tools to make it work. 

So, I decided to see if I could modify [Jesse's ipython code](http://nbviewer.jupyter.org/github/jmsteinw/Notebooks/blob/master/IndeedJobs.ipynb) to get .csv files of the wordcounts of the names of the Data Science Tools from the job ads. His code was written to produce plots of the results right after scraping, not produce .csv files. 

This folder includes the ipython notebooks with the modified codes. 

You may see in the ipython notebook that the code was a bit repetitive. I was trying to stick to d.r.y (don't repeat yourself) principle but I have never programmed in Python, so this will be a work in progress for me to ma